requirements.txt
flask==3.1.*
python-frontmatter==1.1.*
markdown==3.7.*
gunicorn==23.*
requests==2.32.*
python-dotenv==1.0.*
